Compute absolute segment indices via document tree when breaking ties
This replaces the use of big-rational numbers to determine the relative
position of segments when breaking ties during remote insertions and
position calculations.

To avoid repeatedly splaying nodes to the top of the splay tree when
resolving remote positions, we splay the original left and right
dependencies at the start of the tie-breaking algorithm rather than each
time we request and index. Not sure if this is the best approach but it
works.

/cc @as-cii