Include last character when overshooting in `characterIndexForPosition`
This was causing randomized tests to never delete the last character at
the end of a line.

Even though it didn't help find the divergence bug some of our users
have observed, it's interesting to note that `characterIndexForPosition`
is used also when determining where to split a segment. That said, it's
still not obvious that this would cause a problem, because we should
never split after the last character of a segment.