Allocate a new position on each iteration of `characterIndexForPosition`
In trying to debug the divergence problem we have been experiencing, I
noticed that `characterIndexForPosition` exhibited a different behavior
when the debugger was on.

In particular, I was able to observe the assertion error introduced in
0c0c656 when I was not debugging the function. When adding `debugger`
statements or stepping through the source code, the assertion did not
fail. Since v8 disables optimizations when the debugger is on, I started
wondering if this bug was manifesting itself due to some logic error in
the optimizer.

This initially seemed very unlikely and led me to think there was some
other kind of race condition that was being prevented by adding delays
in the execution as I was stepping through the source code and
inspecting variables. However, the assertion stopped failing once this
commit was applied to my local checkout of the real-time package, and I
couldn't seem to reproduce the assertion errors anymore after that
(either with or without the debugger).

I still think this is a pretty mysterious issue, especially because our
randomized tests are not reproducing it at all. However, it's hard to
say whether v8 optimizations are kicking in while running tests, as
every test is wrapped in a try/catch in order to show failure messages.