Streamline DocumentReplica public API and tests
This commit abandons the symmetry of applyLocal/applyRemote in favor of
assuming methods refer to local operations so we can just use positional
arguments. This paves the way for other methods on the replica such as
undo, redo, createCheckpoint, etc.

Instead of the applyLocal and undoLocal methods on DocumentReplica, we
now have insert, delete, and undoOrRedoOperation.

Instead of applyRemote and canApplyRemote, we now have
applyRemoteOperation and canApplyRemoteOperation.

This commit also adds a bunch of test helpers that remove a lot of noise
from the unit tests.