Represent all local changes as patch-style change arrays
Previously, local operations were expressed as individual insertions or
deletions to apply to the buffer. Now, local operations are expressed as
arrays of "change" objects, with the fields oldStart, oldEnd, oldText,
newStart, newEnd, newText. This is the same format that Patch.getChanges
returns and is how the text-buffer library expects the history to
express changes.