Treat restoration of markers via undo/redo separately from replication
Previously, replicated markers were automatically snapshotted when
transactions were created, but this forces us to replicate any markers
that we want to restore across undo and redo. These sets of markers may
be different, as is the case in Atom. With the new API, we take the
current state of markers explicitly when creating and grouping
checkpoints, allowing the caller to choose what markers they want to
restore on undo and redo.