Splay when finding segment containing offset in SplitTree
Previously, we would only splay when splitting a segment. This is
suboptimal, because querying segments can become slower and slower over
time as more segments get added.

This commit splays every time `SplitTree.findSegmentContainingOffset` is
called, ensuring searched segments are always close to the top of the
tree for localized workloads.