Remove obsolete and expensive assertions
We added these back in 0c0c656 to investigate divergence issues, but we
haven't observed any since then and we tracked down the ones we observed
to that mysterious v8 optimization bug (5dbcfb6). Hence, we can safely
remove these assertions and get a nice speedup by avoiding to call
`extentForText` every time we split a segment.