Don't store marker updates in operations history
When replicating a document, `getOperations` will be used instead. This
new method will return:

* All text operations since the creation of the document.
* A synthesized marker update operation for each site that captures the
current state of each marker layer.

This should significantly speed up the replication of any given
document, because the time spent in integrating remote operations into
the new document will only be affected by all the text operations and
the current state of all the markers (as opposed to the entire history
of marker updates since the document was created).